Got error in node_module - Cannot find module 'tesseract.js'.
I got an error while importing the tesseract module in our project module. 
ERROR in node_modules/tesseract.ts/dist/index.d.ts(1,33): error TS2307: Cannot find module 'tesseract.js'.  

I searched in google, some one has suggested to modify the index.d.ts file in node_module in tesseract.ts folder. After doing that change it is working charm. But this is not a right way to do a modification on files in nodule_module. So I would help to fix this issue in this package. Kindly let me know if this is the fix or I'm referring a wrong package.

Thanks